3. I do not fear computers. I fear the lack of them

+++ 0x05-pointers_arrays_strings/main.h
@@ -4,7 +4,7 @@
 int _putchar(char c);
 void reset_to_98(int *n);
 void swap_int(int *a, int *b);
-size_t _strlen(char *str);
+int _strlen(char *str);
 void _puts(char *str);
 void print_rev(char *s);
 void puts2(char *str);
